🚧 8SH7ZM task: apply external agent result
