+++ .agentplane/tasks/202609031902-8SH7ZM/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202609031902-8SH7ZM/repair-plan-amendment-verify-steps-projection-ro",
   "created_at": "2026-09-03T19:09:23.499Z",
-  "diffstat_sha256": "sha256:91141d6182238ffc76f722ed2a2337cfd144d3831188e49bacd79f0e58b38378",
+  "diffstat_sha256": "sha256:447c7435e0662d46fc1cfc3496f515db478e7791822713f4ca88bd7cbe2e7525",
   "last_verified_at": "2026-09-03T19:33:39.803Z",
   "last_verified_diffstat_sha256": "sha256:f6ca54984df83a11a2608aa2a02a64fea3efb9cf78f6e8881ba138003217529e",
   "pr_number": 5891,
